requirements.txt
grandalf==0.8
langchain==1.2.10
langchain_community==0.4.1
langchain-google-genai==4.2.1
langchain-tavily==0.2.17
langgraph==1.0.10
python-dotenv==1.2.1